Contribution: Thanks for qPCR4vir's fix. - "std::size_t-to-unsigned long" issue may crash in 64Bit.

+++ Nana.Cpp11/source/gui/place.cpp
@@ -217,7 +217,7 @@ namespace nana{	namespace gui
 		void _m_throw_error(char err_char)
 		{
 			std::stringstream ss;
-			ss<<"place: invalid character '"<<err_char<<"' at "<<static_cast<std::size_t>(sp_ - divstr_);
+			ss<<"place: invalid character '"<<err_char<<"' at "<<static_cast<unsigned>(sp_ - divstr_);
 			throw std::runtime_error(ss.str());
 		}
 
@@ -247,7 +247,7 @@ namespace nana{	namespace gui
 		void _m_throw_error(const std::string& err)
 		{
 			std::stringstream ss;
-			ss<<"place: "<<err<<" at "<<static_cast<std::size_t>(sp_ - divstr_);
+			ss<<"place: "<<err<<" at "<<static_cast<unsigned>(sp_ - divstr_);
 			throw std::runtime_error(ss.str());
 		}
